% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pkg_chk}
\alias{pkg_chk}
\title{Wrapper for installing and updating the packages I use regularly. Installs dependencies as well.}
\usage{
pkg_chk()
}
\value{
Null
}
\description{
Wrapper for installing and updating the packages I use regularly. Installs dependencies as well.
}
\examples{
pkg_chk()
}
\seealso{
\code{\link{update.packages}}, \code{\link{install.packages}}
}
\keyword{library}
\keyword{update,}

