% Generated by roxygen2 (4.0.2): do not edit by hand
\name{welchEMJ}
\alias{welchEMJ}
\title{Function for calculating Welch's t-test given pairs of means, SDs, and Ns. If vectors are input, the
first element of each of the mean vectors will be compared, then
the second element, and so on.}
\usage{
welchEMJ(m1, m2, s1, s2, n1, n2)
}
\arguments{
\item{m1}{First mean (may be a vector)}

\item{m2}{Second mean}

\item{s1}{First standard deviation}

\item{s2}{Second standard deviation}

\item{n1}{First N}

\item{n2}{Second N}
}
\value{
Data frame containing t, df, and p for each comparison performed.
}
\description{
Function for calculating Welch's t-test given pairs of means, SDs, and Ns. If vectors are input, the
first element of each of the mean vectors will be compared, then
the second element, and so on.
}
\examples{
welchEMJ(0,1,1,2,100,150) # Single comparison
welchEMJ(c(0,0),c(0.5,1),c(1,1),c(2,2),c(100,100),c(150,150)) # Two comparisons
}
\seealso{
\code{\link{CourseEvalTable}}
}
\keyword{Welch,}
\keyword{t-test}

